+++ packages/bridge-status-controller/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Bump `@metamask/bridge-controller` from `^64.4.1` to `^^64.5.0` ([#7622](https://github.com/MetaMask/core/pull/7622))
+
 ## [64.4.2]
 
 ### Changed